Implemented unit test for ONSResponse
